Avoid copying when capacity is readily available.
PCMPISTRI scans 16 bytes at a time so goj relies on padding for perf. It turns out that if there is already 16bytes padding, we can use it rather than copy.
So when we detect the spare capacity, we fill it with zeroes instead of copying. If there's not enough spare capacity, fallback to previous behavior.